+++ infra/growth/src/scripts/createProdCampaign.js
@@ -201,6 +201,273 @@ const aprRules = {
   }
 }
 
+const mayRules = {
+  numLevels: 3,
+  levels: {
+    0: {
+      rules: [
+        {
+          id: 'ProfilePublished',
+          class: 'SingleEvent',
+          config: {
+            eventType: 'ProfilePublished',
+            reward: null,
+            nextLevelCondition: false,
+            visible: true,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'EmailAttestation',
+          class: 'SingleEvent',
+          config: {
+            eventType: 'EmailAttestationPublished',
+            reward: null,
+            visible: true,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'BothRules',
+          class: 'MultiEvents',
+          config: {
+            eventTypes: ['ProfilePublished', 'EmailAttestationPublished'],
+            visible: false,
+            numEventsRequired: 2,
+            reward: null,
+            nextLevelCondition: true,
+            scope: 'user',
+            unlockConditionMsg: [
+              {
+                conditionTranslateKey: 'growth.profile.requirement',
+                conditionIcon: 'images/growth/email-icon-small.svg'
+              },
+              {
+                conditionTranslateKey: 'growth.email.requirement',
+                conditionIcon: 'images/growth/email-icon-small.svg'
+              }
+            ]
+          }
+        }
+      ]
+    },
+    1: {
+      rules: [
+        {
+          id: 'PhoneAttestation',
+          class: 'SingleEvent',
+          config: {
+            eventType: 'PhoneAttestationPublished',
+            reward: {
+              amount: tokenNaturalUnits(25),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 1,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'FacebookAttestation',
+          class: 'SingleEvent',
+          config: {
+            eventType: 'FacebookAttestationPublished',
+            reward: {
+              amount: tokenNaturalUnits(25),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 1,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'AirbnbAttestation',
+          class: 'SingleEvent',
+          config: {
+            eventType: 'AirbnbAttestationPublished',
+            reward: {
+              amount: tokenNaturalUnits(25),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 1,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'TwitterAttestation',
+          class: 'SingleEvent',
+          config: {
+            eventType: 'TwitterAttestationPublished',
+            reward: {
+              amount: tokenNaturalUnits(25),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 1,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'GoogleAttestation',
+          class: 'SingleEvent',
+          config: {
+            eventType: 'GoogleAttestationPublished',
+            reward: {
+              amount: tokenNaturalUnits(25),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 1,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'TwoAttestations',
+          class: 'MultiEvents',
+          config: {
+            eventTypes: [
+              'PhoneAttestationPublished',
+              'FacebookAttestationPublished',
+              'AirbnbAttestationPublished',
+              'TwitterAttestationPublished',
+              'GoogleAttestationPublished'
+            ],
+            visible: false,
+            numEventsRequired: 2,
+            reward: null,
+            nextLevelCondition: true,
+            scope: 'user',
+            unlockConditionMsg: [
+              {
+                conditionTranslateKey: 'growth.two.attestations.requirement',
+                conditionIcon: 'images/growth/attestation-icon.svg'
+              }
+            ]
+          }
+        }
+      ]
+    },
+    2: {
+      rules: [
+        {
+          id: 'Referral',
+          class: 'Referral',
+          config: {
+            levelRequired: 2,
+            reward: {
+              amount: tokenNaturalUnits(50),
+              currency: 'OGN'
+            },
+            limit: 25,
+            visible: true,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'ListingPurchaseTShirt',
+          class: 'ListingPurchase',
+          config: {
+            eventType: 'ListingPurchased',
+            listingId: '1-000-1',
+            reward: {
+              amount: tokenNaturalUnits(25),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 100,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'ListingPurchaseGC',
+          class: 'ListingPurchase',
+          config: {
+            listingId: '1-000-2',
+            reward: {
+              amount: tokenNaturalUnits(50),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 100,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'ListingPurchaseDonation',
+          class: 'ListingPurchase',
+          config: {
+            listingId: '1-000-3',
+            reward: {
+              amount: tokenNaturalUnits(500),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 100,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'ListingPurchaseHousing',
+          class: 'ListingPurchase',
+          config: {
+            listingId: '1-000-4',
+            reward: {
+              amount: tokenNaturalUnits(100),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 100,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'ListingPurchaseInfluencer',
+          class: 'ListingPurchase',
+          config: {
+            listingId: '1-000-5',
+            reward: {
+              amount: tokenNaturalUnits(100),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 100,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        },
+        {
+          id: 'ListingPurchaseArt',
+          class: 'ListingPurchase',
+          config: {
+            listingId: '1-000-6',
+            reward: {
+              amount: tokenNaturalUnits(1000),
+              currency: 'OGN'
+            },
+            visible: true,
+            limit: 100,
+            nextLevelCondition: false,
+            scope: 'campaign'
+          }
+        }
+      ]
+    }
+  }
+}
+
 async function createAprilProdCampaign() {
   console.log('Creating April campaign data in prod...')
 
@@ -221,7 +488,32 @@ async function createAprilProdCampaign() {
   })
 }
 
-createAprilProdCampaign().then(() => {
+async function createMayProdCampaign() {
+  console.log('Creating May campaign data in prod...')
+
+  /* IMPORTANT when adding new translatable fields update the enums document:
+   * origin-dapp/src/constants/Growth$FbtEnum.js
+   */
+  await db.GrowthCampaign.create({
+    nameKey: 'growth.may2019.name',
+    shortNameKey: 'growth.may2019.short_name',
+    rules: JSON.stringify(mayRules),
+    startDate: Date.parse('May 1, 2019, 00:00 UTC'),
+    endDate: Date.parse('June 1, 2019, 00:00 UTC'),
+    distributionDate: Date.parse('June 1, 2019, 00:00 UTC'),
+    cap: tokenNaturalUnits(1000000), // Set cap to 1M tokens
+    capUsed: 0,
+    currency: 'OGN',
+    rewardStatus: enums.GrowthCampaignRewardStatuses.Active
+  })
+}
+
+//createAprilProdCampaign().then(() => {
+//  console.log('Done')
+//  process.exit()
+//})
+
+createMayProdCampaign().then(() => {
   console.log('Done')
   process.exit()
 })
